+++ changelog/unreleased/pr-16344.toml
@@ -1,4 +1,4 @@
-type = "c"
+type = "a"
 message = "Add change field type action for a field"
 
 pulls = ["16344"]
